+++ AoE2ScenarioRms/enums/internal/xs_key.py
@@ -19,6 +19,7 @@ class XsKey(Enum):
     CONFIG_DECLARATION = auto()
 
     AFTER_RESOURCE_SPAWN_EVENT = auto()
+    AFTER_ALL_RESOURCES_SPAWNED_EVENT = auto()
 
     XS_ON_INIT_FILE = auto()
     XS_ON_INIT_RULE = auto()
@@ -30,21 +31,23 @@ class XsKey(Enum):
         return _xs_join_strings[key]
 
 
+_tab = ' ' * 4
 _xs_join_strings = {
     XsKey.RESOURCE_VARIABLE_COUNT: '',
 
     XsKey.RESOURCE_VARIABLE_DECLARATION: '\n',
     XsKey.XS_ON_INIT_FILE: '\n',
 
-    XsKey.RESOURCE_GROUP_NAMES_DECLARATION: '\n\t',
-    XsKey.RESOURCE_COUNT_DECLARATION: '\n\t',
-    XsKey.RESOURCE_MAX_SPAWN_DECLARATION: '\n\t',
-    XsKey.RESOURCE_MAX_SPAWN_IS_PER_PLAYER_DECLARATION: '\n\t',
-    XsKey.RESOURCE_LOCATION_INJECTION: '\n\t',
-    XsKey.CONFIG_DECLARATION: '\n\t',
-    XsKey.XS_ON_INIT_RULE: '\n\t',
+    XsKey.RESOURCE_GROUP_NAMES_DECLARATION: f'\n{_tab}',
+    XsKey.RESOURCE_COUNT_DECLARATION: f'\n{_tab}',
+    XsKey.RESOURCE_MAX_SPAWN_DECLARATION: f'\n{_tab}',
+    XsKey.RESOURCE_MAX_SPAWN_IS_PER_PLAYER_DECLARATION: f'\n{_tab}',
+    XsKey.RESOURCE_LOCATION_INJECTION: f'\n{_tab}',
+    XsKey.CONFIG_DECLARATION: f'\n{_tab}',
+    XsKey.XS_ON_INIT_RULE: f'\n{_tab}',
 
-    XsKey.AFTER_RESOURCE_SPAWN_EVENT: '\n\t\t',
+    XsKey.AFTER_RESOURCE_SPAWN_EVENT: f'\n{_tab}',
+    XsKey.AFTER_ALL_RESOURCES_SPAWNED_EVENT: f'\n{_tab}{_tab}',
 
-    XsKey.XS_ON_SUCCESSFUL_SPAWN: '\n\t\t\t',
+    XsKey.XS_ON_SUCCESSFUL_SPAWN: f'\n{_tab}{_tab}{_tab}',
 }